Massive work on workspace logic.  Christ.
Fixed json polymorphic typing.  Use @type instead of "type" for the type
of the json node.  We use "type" as an instance variable name, and this
collides.  Not sure how this ever worked ;)

Seriously cleaned up the bootstrap.  Add ability to add to a workspace,
rather than just create a workspace from workspace dsl.  Use the new
kernel.wsp to bootstrap kernel facets and the kernel workspace (as a
facet).  Geebus.  Been meaning to do this for a while.

Fix frontier serialization in workspaces.  That has an ugly hack because
we use null for values that ain't.  Need to fix that in a more clear
way.

Fixed default values for authorized attributes, taking into account
network edge attributes.

rationalize workspace dsl package.