New update mutation encoding.
Generate an update input type, use that as the "state" argument to the
UpdateX mutation.  This solves my problem with an exploding mutation
field namespace and sure seems like a natural epression of this.  I'll
also be able to add the "add" "remove" "getImediate" phantasm mutators
in a very obvious and compact fashion, using addition fields.