Manifest kernel workspace Phantasms.
Move as much as I dare to the kernel.wsp DSL.  Also took the opportunity
to model integration of JVM based phantasm implementations.  Workspaces
now have Plugins which map to defined facets in the workspace.  These
plugins can define constructors, instance/static methods.  Constructors
adn instance/static methods have an implementation class that maps to
some class that can be resolved from the class scope.  The class scope
is defined to be the current thread's class loader + the scope defined
by the Code Sources of the Plugin.

Constructor and instance/static methods also have an attribute that
defines their implementation method name.  This maps to a static method
name on the implementation class.

instance/static methods have an attribute defining the return type of
the method.  This will map to a graphQL type, resolved in the generated
facet types of the workspace.  instance/static methods also have input
arguments.  These input arguments are also graphQL types, resolved in
the workspace.

Haven't defined the formal interface to the plugin, but that will come
next.