version workspaces.  workspace importer obeys "Liquibase semantics"
workspaces will be created if version is 0.  If workspace exists in DB,
this will throw an illegal state exception telling you that.  For non 0
versions, this is treated as an incremental load and will throw an
illegal state if the current version of the workspace in the DB is not
(version - 1).