lot of clean up on graph slicing and traversal.
Normalized the traversal process - as much as I can, anyway.  We now
create a clone of the ruleform, and null out any FK references for
serializing frontier exits.  This solves the stupid problem with
attribute value translation due to a required indirection to the
attribute for the type.  Geebus, that was hard to track down.

Delta graph production looks to be working (finally).  Adding tests...