Capabilities are now OR'd across principal and active roles.
This allows dynamic role capabilities.  The AuthorizedPrinicpal has the
authenticated principal and a set of roles that asserted as active.
These roles are facets of the authenticated principal. 

To get dynamic capabilities, we OR the capability checks on the
principal and the asserted role classifications.  This solves some
rather tricky security use cases cleanly and with low cognitive overhead