Add updated timestamp on all attribute values.
This was missing and needed to get in before the big schema freeze.
This adds back the time stamp when the attribute value was updated that
we used to have before versioning supplanted that for optimistic
locking.  It's something we would like to know anyway for these
primitive values.

It also solves my problem with the capability based, dynamic roles and
the OAuth bearer token.  Previously, the bearer access token attribute
was a timstamp value.  I need to store additional attribues in addition
to the last updated timestamp and correlation between these related
attributes was looking to be a nightmare.  So, by doing this, I can
store things like IP address, etc, as JSON in the attribute value and
use the update timestamp as I was previously using the value.

What this means is that we now have fully blown dynamic capabilities we
can pass around by a UUID.  Timeout check is a single DB lookup of a
single row's column value.  The value of the access token is the JSON
credential created at login time.

Fucking awesome, if you ask me.  And naturally falls out of modeling
this in the Ultrastructure way.