Dynamic capabilities. Fuck yea.  Shit's getting real.
Round trip validatation of AuthX resource and bearer token
authenticator.  Store remote IP and capabilities in AccessToken
attribute on the CoreUser.  AuthorizedPrincipal populated from the
Credential stored in the AccessToken value.  Asserted facets that are
implicitly or inferered on the principal are set on the
AuthorizedPrincipal by the agency bearer token authenticator.