state snapshots, state admin commands for HandiNavi
State snapshots are defined to be the state of the CoRE instance
subtracting out the instance state that is defined by a workspace
authorization.  That is, the set of ruleforms in the instance where the
workspace_authorization column is null.

Added commands to load/dump snapshot state.  Added command to load
workspace snapshots.  Renamed some commands for clarity.