centralize layout from CSS.
Lot's of rework, as we now apply the layout constants differently.
Realized that the relation columns are immaterial to the layout except
in propagating width.

Do the layout earlier than waiting for the layout to occur so I don't
see a lot of updates.  Scroll bars are linked after layout.